requirements.txt
python_fasthtml==0.12.1